Moved the UIViewController extension to its own file
